Fix : 각종 코드 정리 및 오류 수정
