+++ src/main/java/controller/RequestMapping.java
@@ -10,7 +10,6 @@ import controller.group.*;
 import controller.challenge.*;
 import controller.todo.*;
 
-
 public class RequestMapping {
     private static final Logger logger = LoggerFactory.getLogger(DispatcherServlet.class);
