+++ src/main/java/controller/user/LogoutController.java
@@ -9,7 +9,7 @@ import controller.Controller;
 public class LogoutController implements Controller {
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
-		//세션에 저장된 사용자 이이디를 삭제하고 세션을 무효화 함 
+		
 		HttpSession session = request.getSession();
 		session.removeAttribute(UserSessionUtils.USER_SESSION_KEY);
 		session.invalidate();
