+++ src/main/java/model/dao/mybatis/mapper/HbtiMapper.java
@@ -1,7 +1,6 @@
 package model.dao.mybatis.mapper;
 
 import java.util.List;
-import java.util.Map;
 
 import model.HBTI;
 
@@ -10,6 +9,8 @@ public interface HbtiMapper {
 
 	public List<Integer> selectGroupByHbtiId(int hbti_id);
 
+	public String selectHbtiImgByName(String name);
+	
 	public int selectGroupMemberNumByGroupId(int group_id);
 
 	public List<String> selectUserListByGroupId(int group_id);
