+++ src/main/java/model/dao/mybatis/mapper/UserMapper.java
@@ -34,5 +34,4 @@ public interface UserMapper {
 	
 	public List<String> selectChallengePostListByUserId(String user_id);
 	
-
 }
