+++ src/main/resources/model/dao/mybatis/mapper/UserMapper.xml
@@ -114,8 +114,4 @@
 		WHERE writer_id = #{user_id}
 	</select>
 
-
-
-
-
 </mapper>
