+++ src/main/webapp/css/user/hbtiTestForm.css
@@ -46,6 +46,7 @@
 	font-style: normal;
 }
 
+/*main*/
 body {
 	width: 100%;
 	background-color: #112D4E;
