+++ src/main/webapp/css/user/joinForm.css
@@ -59,6 +59,7 @@
 	font-style: normal;
 }
 
+/*main*/
 body {
 	width: 100%;
 	background-color: #112D4E;
