+++ src/main/webapp/css/user/loginForm.css
@@ -138,7 +138,7 @@ td {
 	border-radius: 500px;
 }
 
-/* login form */
+/* login_form */
 form {
 	padding: 150px 0 0 0;
 }
