+++ src/main/webapp/css/user/myPage.css
@@ -96,8 +96,6 @@ img {
 	align-items: center;
 }
 
-
-
 #info-name, #info-HBTI, #group-name {
 	font-size: 20px;
 }
